Calculating per transaction tax obligation (#2)
Each transaction has withhold tax, sometimes it may be equal zero.
For polish taxation obligation each transaction generates 19% tax obligation, but the withhold tax is deducted from that. What is left after the deduction is the final obligation in Poland. Sometimes, due to withhold tax being paid in foreign currency and rounding, after exchange recalculation it may happen that withhold tax is higher than 19% obligation in Poland - in such case there is nothing to deduct.

Next step will be to sum up all the remaining taxes to be paid and provide clear information regarding what needs to be filled in PIT-38 form (and hopefully where it needs to be filled in).